+++ gems/quilt_rails/app/controllers/quilt/performance_report_controller.rb
@@ -5,12 +5,15 @@ module Quilt
     include Quilt::Performance::Reportable
     protect_from_forgery with: :null_session
 
-    def create
-      process_report
+    MAX_THREADS = 25
 
-      render(json: { result: 'success' }, status: 200)
-    rescue ActionController::ParameterMissing => error
-      render(json: { error: error.message, status: 422 })
+    def create
+      Thread.new do
+        process_report
+      rescue # rubocop:disable Lint/SuppressedException
+      end if Thread.list.count < MAX_THREADS
+
+      render(plain: '{"result":"success"}', status: 200)
     end
   end
 end